requirements.txt
Flask==2.2.2
PyJWT==2.6.0
pyodbc==4.0.35
python-hostlist==1.23.0
netaddr==0.8.0
cryptography==38.0.4
gunicorn==20.1.0
libtorrent==2.0.9
python3-snmp==1.1a1
